Struct aws_sdk_storagegateway::client::fluent_builders::UpdateNFSFileShare

pub struct UpdateNFSFileShare<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> UpdateNFSFileShare<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(
self
) -> Result<UpdateNfsFileShareOutput, SdkError<UpdateNFSFileShareError>> where
R::Policy: SmithyRetryPolicy<UpdateNfsFileShareInputOperationOutputAlias, UpdateNfsFileShareOutput, UpdateNFSFileShareError, UpdateNfsFileShareInputOperationRetryAlias>,

pub fn file_share_arn(self, inp: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the file share to be updated.

pub fn set_file_share_arn(self, input: Option<String>) -> Self

pub fn kms_encrypted(self, inp: bool) -> Self

Set to true to use Amazon S3 server-side encryption with your own KMS key, or false to use a key managed by Amazon S3. Optional.

Valid Values: true | false

pub fn set_kms_encrypted(self, input: Option<bool>) -> Self

pub fn kms_key(self, inp: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of a symmetric customer master key (CMK) used for Amazon S3 server-side encryption. Storage Gateway does not support asymmetric CMKs. This value can only be set when KMSEncrypted is true. Optional.

pub fn set_kms_key(self, input: Option<String>) -> Self

pub fn nfs_file_share_defaults(self, inp: NfsFileShareDefaults) -> Self

The default values for the file share. Optional.

pub fn set_nfs_file_share_defaults(
self,
input: Option<NfsFileShareDefaults>
) -> Self

pub fn default_storage_class(self, inp: impl Into<String>) -> Self

The default storage class for objects put into an Amazon S3 bucket by the S3 File Gateway. The default value is S3_INTELLIGENT_TIERING. Optional.

Valid Values: S3_STANDARD | S3_INTELLIGENT_TIERING | S3_STANDARD_IA | S3_ONEZONE_IA

pub fn set_default_storage_class(self, input: Option<String>) -> Self

pub fn object_acl(self, inp: ObjectAcl) -> Self

A value that sets the access control list (ACL) permission for objects in the S3 bucket that a S3 File Gateway puts objects into. The default value is private.

pub fn set_object_acl(self, input: Option<ObjectAcl>) -> Self

pub fn client_list(self, inp: impl Into<String>) -> Self

Appends an item to ClientList.

To override the contents of this collection use set_client_list.

The list of clients that are allowed to access the S3 File Gateway. The list must contain either valid IP addresses or valid CIDR blocks.

pub fn set_client_list(self, input: Option<Vec<String>>) -> Self

pub fn squash(self, inp: impl Into<String>) -> Self

The user mapped to anonymous user.

Valid values are the following:

RootSquash: Only root is mapped to anonymous user.

NoSquash: No one is mapped to anonymous user.

AllSquash: Everyone is mapped to anonymous user.

pub fn set_squash(self, input: Option<String>) -> Self

pub fn read_only(self, inp: bool) -> Self

A value that sets the write status of a file share. Set this value to true to set the write status to read-only, otherwise set to false.

Valid Values: true | false

pub fn set_read_only(self, input: Option<bool>) -> Self

pub fn guess_mime_type_enabled(self, inp: bool) -> Self

A value that enables guessing of the MIME type for uploaded objects based on file extensions. Set this value to true to enable MIME type guessing, otherwise set to false. The default value is true.

Valid Values: true | false

pub fn set_guess_mime_type_enabled(self, input: Option<bool>) -> Self

pub fn requester_pays(self, inp: bool) -> Self

A value that sets who pays the cost of the request and the cost associated with data download from the S3 bucket. If this value is set to true, the requester pays the costs; otherwise, the S3 bucket owner pays. However, the S3 bucket owner always pays the cost of storing data.

RequesterPays is a configuration for the S3 bucket that backs the file share, so make sure that the configuration on the file share is the same as the S3 bucket configuration.

Valid Values: true | false

pub fn set_requester_pays(self, input: Option<bool>) -> Self

pub fn file_share_name(self, inp: impl Into<String>) -> Self

The name of the file share. Optional.

FileShareName must be set if an S3 prefix name is set in LocationARN.

pub fn set_file_share_name(self, input: Option<String>) -> Self

pub fn cache_attributes(self, inp: CacheAttributes) -> Self

specifies refresh cache information for the file share.

pub fn set_cache_attributes(self, input: Option<CacheAttributes>) -> Self

pub fn notification_policy(self, inp: impl Into<String>) -> Self

The notification policy of the file share. SettlingTimeInSeconds controls the number of seconds to wait after the last point in time a client wrote to a file before generating an ObjectUploaded notification. Because clients can make many small writes to files, it's best to set this parameter for as long as possible to avoid generating multiple notifications for the same file in a small time period.

SettlingTimeInSeconds has no effect on the timing of the object uploading to Amazon S3, only the timing of the notification.

The following example sets NotificationPolicy on with SettlingTimeInSeconds set to 60.

{\"Upload\": {\"SettlingTimeInSeconds\": 60}}

The following example sets NotificationPolicy off.

{}

pub fn set_notification_policy(self, input: Option<String>) -> Self