Struct aws_sdk_storagegateway::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for StorageGateway_20130630.

This client allows ergonomic access to a StorageGateway_20130630-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

pub fn conf(&self) -> &Config

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn activate_gateway(&self) -> ActivateGateway<C, M, R>

pub fn add_cache(&self) -> AddCache<C, M, R>

pub fn add_tags_to_resource(&self) -> AddTagsToResource<C, M, R>

pub fn add_upload_buffer(&self) -> AddUploadBuffer<C, M, R>

pub fn add_working_storage(&self) -> AddWorkingStorage<C, M, R>

pub fn assign_tape_pool(&self) -> AssignTapePool<C, M, R>

pub fn associate_file_system(&self) -> AssociateFileSystem<C, M, R>

pub fn attach_volume(&self) -> AttachVolume<C, M, R>

pub fn cancel_archival(&self) -> CancelArchival<C, M, R>

pub fn cancel_retrieval(&self) -> CancelRetrieval<C, M, R>

pub fn create_cachedi_scsi_volume(&self) -> CreateCachediSCSIVolume<C, M, R>

pub fn create_nfs_file_share(&self) -> CreateNFSFileShare<C, M, R>

pub fn create_smb_file_share(&self) -> CreateSMBFileShare<C, M, R>

pub fn create_snapshot(&self) -> CreateSnapshot<C, M, R>

pub fn create_snapshot_from_volume_recovery_point(
&self
) -> CreateSnapshotFromVolumeRecoveryPoint<C, M, R>

pub fn create_storedi_scsi_volume(&self) -> CreateStorediSCSIVolume<C, M, R>

pub fn create_tape_pool(&self) -> CreateTapePool<C, M, R>

pub fn create_tapes(&self) -> CreateTapes<C, M, R>

pub fn create_tape_with_barcode(&self) -> CreateTapeWithBarcode<C, M, R>

pub fn delete_automatic_tape_creation_policy(
&self
) -> DeleteAutomaticTapeCreationPolicy<C, M, R>

pub fn delete_bandwidth_rate_limit(&self) -> DeleteBandwidthRateLimit<C, M, R>

pub fn delete_chap_credentials(&self) -> DeleteChapCredentials<C, M, R>

pub fn delete_file_share(&self) -> DeleteFileShare<C, M, R>

pub fn delete_gateway(&self) -> DeleteGateway<C, M, R>

pub fn delete_snapshot_schedule(&self) -> DeleteSnapshotSchedule<C, M, R>

pub fn delete_tape(&self) -> DeleteTape<C, M, R>

pub fn delete_tape_archive(&self) -> DeleteTapeArchive<C, M, R>

pub fn delete_tape_pool(&self) -> DeleteTapePool<C, M, R>

pub fn delete_volume(&self) -> DeleteVolume<C, M, R>

pub fn describe_availability_monitor_test(
&self
) -> DescribeAvailabilityMonitorTest<C, M, R>

pub fn describe_bandwidth_rate_limit(
&self
) -> DescribeBandwidthRateLimit<C, M, R>

pub fn describe_bandwidth_rate_limit_schedule(
&self
) -> DescribeBandwidthRateLimitSchedule<C, M, R>

pub fn describe_cache(&self) -> DescribeCache<C, M, R>

pub fn describe_cachedi_scsi_volumes(
&self
) -> DescribeCachediSCSIVolumes<C, M, R>

pub fn describe_chap_credentials(&self) -> DescribeChapCredentials<C, M, R>

pub fn describe_file_system_associations(
&self
) -> DescribeFileSystemAssociations<C, M, R>

pub fn describe_gateway_information(
&self
) -> DescribeGatewayInformation<C, M, R>

pub fn describe_maintenance_start_time(
&self
) -> DescribeMaintenanceStartTime<C, M, R>

pub fn describe_nfs_file_shares(&self) -> DescribeNFSFileShares<C, M, R>

pub fn describe_smb_file_shares(&self) -> DescribeSMBFileShares<C, M, R>

pub fn describe_smb_settings(&self) -> DescribeSMBSettings<C, M, R>

pub fn describe_snapshot_schedule(&self) -> DescribeSnapshotSchedule<C, M, R>

pub fn describe_storedi_scsi_volumes(
&self
) -> DescribeStorediSCSIVolumes<C, M, R>

pub fn describe_tape_archives(&self) -> DescribeTapeArchives<C, M, R>

pub fn describe_tape_recovery_points(
&self
) -> DescribeTapeRecoveryPoints<C, M, R>

pub fn describe_tapes(&self) -> DescribeTapes<C, M, R>

pub fn describe_upload_buffer(&self) -> DescribeUploadBuffer<C, M, R>

pub fn describe_vtl_devices(&self) -> DescribeVTLDevices<C, M, R>

pub fn describe_working_storage(&self) -> DescribeWorkingStorage<C, M, R>

pub fn detach_volume(&self) -> DetachVolume<C, M, R>

pub fn disable_gateway(&self) -> DisableGateway<C, M, R>

pub fn disassociate_file_system(&self) -> DisassociateFileSystem<C, M, R>

pub fn join_domain(&self) -> JoinDomain<C, M, R>

pub fn list_automatic_tape_creation_policies(
&self
) -> ListAutomaticTapeCreationPolicies<C, M, R>

pub fn list_file_shares(&self) -> ListFileShares<C, M, R>

pub fn list_file_system_associations(
&self
) -> ListFileSystemAssociations<C, M, R>

pub fn list_gateways(&self) -> ListGateways<C, M, R>

pub fn list_local_disks(&self) -> ListLocalDisks<C, M, R>

pub fn list_tags_for_resource(&self) -> ListTagsForResource<C, M, R>

pub fn list_tape_pools(&self) -> ListTapePools<C, M, R>

pub fn list_tapes(&self) -> ListTapes<C, M, R>

pub fn list_volume_initiators(&self) -> ListVolumeInitiators<C, M, R>

pub fn list_volume_recovery_points(&self) -> ListVolumeRecoveryPoints<C, M, R>

pub fn list_volumes(&self) -> ListVolumes<C, M, R>

pub fn notify_when_uploaded(&self) -> NotifyWhenUploaded<C, M, R>

pub fn refresh_cache(&self) -> RefreshCache<C, M, R>

pub fn remove_tags_from_resource(&self) -> RemoveTagsFromResource<C, M, R>

pub fn reset_cache(&self) -> ResetCache<C, M, R>

pub fn retrieve_tape_archive(&self) -> RetrieveTapeArchive<C, M, R>

pub fn retrieve_tape_recovery_point(&self) -> RetrieveTapeRecoveryPoint<C, M, R>

pub fn set_local_console_password(&self) -> SetLocalConsolePassword<C, M, R>

pub fn set_smb_guest_password(&self) -> SetSMBGuestPassword<C, M, R>

pub fn shutdown_gateway(&self) -> ShutdownGateway<C, M, R>

pub fn start_availability_monitor_test(
&self
) -> StartAvailabilityMonitorTest<C, M, R>

pub fn start_gateway(&self) -> StartGateway<C, M, R>

pub fn update_automatic_tape_creation_policy(
&self
) -> UpdateAutomaticTapeCreationPolicy<C, M, R>

pub fn update_bandwidth_rate_limit(&self) -> UpdateBandwidthRateLimit<C, M, R>

pub fn update_bandwidth_rate_limit_schedule(
&self
) -> UpdateBandwidthRateLimitSchedule<C, M, R>

pub fn update_chap_credentials(&self) -> UpdateChapCredentials<C, M, R>

pub fn update_file_system_association(
&self
) -> UpdateFileSystemAssociation<C, M, R>

pub fn update_gateway_information(&self) -> UpdateGatewayInformation<C, M, R>

pub fn update_gateway_software_now(&self) -> UpdateGatewaySoftwareNow<C, M, R>

pub fn update_maintenance_start_time(
&self
) -> UpdateMaintenanceStartTime<C, M, R>

pub fn update_nfs_file_share(&self) -> UpdateNFSFileShare<C, M, R>

pub fn update_smb_file_share(&self) -> UpdateSMBFileShare<C, M, R>

pub fn update_smb_file_share_visibility(
&self
) -> UpdateSMBFileShareVisibility<C, M, R>

pub fn update_smb_security_strategy(&self) -> UpdateSMBSecurityStrategy<C, M, R>

pub fn update_snapshot_schedule(&self) -> UpdateSnapshotSchedule<C, M, R>

pub fn update_vtl_device_type(&self) -> UpdateVTLDeviceType<C, M, R>

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

pub fn from_conf(conf: Config) -> Self