Module aws_sdk_storagegateway::error

Modules

internal_server_error

See InternalServerError

invalid_gateway_request_exception

See InvalidGatewayRequestException

service_unavailable_error

See ServiceUnavailableError

Structs

ActivateGatewayError

AddCacheError

AddTagsToResourceError

AddUploadBufferError

AddWorkingStorageError

AssignTapePoolError

AssociateFileSystemError

AttachVolumeError

CancelArchivalError

CancelRetrievalError

CreateCachediSCSIVolumeError

CreateNFSFileShareError

CreateSMBFileShareError

CreateSnapshotError

CreateSnapshotFromVolumeRecoveryPointError

CreateStorediSCSIVolumeError

CreateTapePoolError

CreateTapeWithBarcodeError

CreateTapesError

DeleteAutomaticTapeCreationPolicyError

DeleteBandwidthRateLimitError

DeleteChapCredentialsError

DeleteFileShareError

DeleteGatewayError

DeleteSnapshotScheduleError

DeleteTapeArchiveError

DeleteTapeError

DeleteTapePoolError

DeleteVolumeError

DescribeAvailabilityMonitorTestError

DescribeBandwidthRateLimitError

DescribeBandwidthRateLimitScheduleError

DescribeCacheError

DescribeCachediSCSIVolumesError

DescribeChapCredentialsError

DescribeFileSystemAssociationsError

DescribeGatewayInformationError

DescribeMaintenanceStartTimeError

DescribeNFSFileSharesError

DescribeSMBFileSharesError

DescribeSMBSettingsError

DescribeSnapshotScheduleError

DescribeStorediSCSIVolumesError

DescribeTapeArchivesError

DescribeTapeRecoveryPointsError

DescribeTapesError

DescribeUploadBufferError

DescribeVTLDevicesError

DescribeWorkingStorageError

DetachVolumeError

DisableGatewayError

DisassociateFileSystemError

InternalServerError

An internal server error has occurred during the request. For more information, see the

InvalidGatewayRequestException

An exception occurred because an invalid gateway request was issued to the service. For

JoinDomainError

ListAutomaticTapeCreationPoliciesError

ListFileSharesError

ListFileSystemAssociationsError

ListGatewaysError

ListLocalDisksError

ListTagsForResourceError

ListTapePoolsError

ListTapesError

ListVolumeInitiatorsError

ListVolumeRecoveryPointsError

ListVolumesError

NotifyWhenUploadedError

RefreshCacheError

RemoveTagsFromResourceError

ResetCacheError

RetrieveTapeArchiveError

RetrieveTapeRecoveryPointError

ServiceUnavailableError

An internal server error has occurred because the service is unavailable. For more

SetLocalConsolePasswordError

SetSMBGuestPasswordError

ShutdownGatewayError

StartAvailabilityMonitorTestError

StartGatewayError

UpdateAutomaticTapeCreationPolicyError

UpdateBandwidthRateLimitError

UpdateBandwidthRateLimitScheduleError

UpdateChapCredentialsError

UpdateFileSystemAssociationError

UpdateGatewayInformationError

UpdateGatewaySoftwareNowError

UpdateMaintenanceStartTimeError

UpdateNFSFileShareError

UpdateSMBFileShareError

UpdateSMBFileShareVisibilityError

UpdateSMBSecurityStrategyError

UpdateSnapshotScheduleError

UpdateVTLDeviceTypeError

Enums

ActivateGatewayErrorKind

AddCacheErrorKind

AddTagsToResourceErrorKind

AddUploadBufferErrorKind

AddWorkingStorageErrorKind

AssignTapePoolErrorKind

AssociateFileSystemErrorKind

AttachVolumeErrorKind

CancelArchivalErrorKind

CancelRetrievalErrorKind

CreateCachediSCSIVolumeErrorKind

CreateNFSFileShareErrorKind

CreateSMBFileShareErrorKind

CreateSnapshotErrorKind

CreateSnapshotFromVolumeRecoveryPointErrorKind

CreateStorediSCSIVolumeErrorKind

CreateTapePoolErrorKind

CreateTapeWithBarcodeErrorKind

CreateTapesErrorKind

DeleteAutomaticTapeCreationPolicyErrorKind

DeleteBandwidthRateLimitErrorKind

DeleteChapCredentialsErrorKind

DeleteFileShareErrorKind

DeleteGatewayErrorKind

DeleteSnapshotScheduleErrorKind

DeleteTapeArchiveErrorKind

DeleteTapeErrorKind

DeleteTapePoolErrorKind

DeleteVolumeErrorKind

DescribeAvailabilityMonitorTestErrorKind

DescribeBandwidthRateLimitErrorKind

DescribeBandwidthRateLimitScheduleErrorKind

DescribeCacheErrorKind

DescribeCachediSCSIVolumesErrorKind

DescribeChapCredentialsErrorKind

DescribeFileSystemAssociationsErrorKind

DescribeGatewayInformationErrorKind

DescribeMaintenanceStartTimeErrorKind

DescribeNFSFileSharesErrorKind

DescribeSMBFileSharesErrorKind

DescribeSMBSettingsErrorKind

DescribeSnapshotScheduleErrorKind

DescribeStorediSCSIVolumesErrorKind

DescribeTapeArchivesErrorKind

DescribeTapeRecoveryPointsErrorKind

DescribeTapesErrorKind

DescribeUploadBufferErrorKind

DescribeVTLDevicesErrorKind

DescribeWorkingStorageErrorKind

DetachVolumeErrorKind

DisableGatewayErrorKind

DisassociateFileSystemErrorKind

JoinDomainErrorKind

ListAutomaticTapeCreationPoliciesErrorKind

ListFileSharesErrorKind

ListFileSystemAssociationsErrorKind

ListGatewaysErrorKind

ListLocalDisksErrorKind

ListTagsForResourceErrorKind

ListTapePoolsErrorKind

ListTapesErrorKind

ListVolumeInitiatorsErrorKind

ListVolumeRecoveryPointsErrorKind

ListVolumesErrorKind

NotifyWhenUploadedErrorKind

RefreshCacheErrorKind

RemoveTagsFromResourceErrorKind

ResetCacheErrorKind

RetrieveTapeArchiveErrorKind

RetrieveTapeRecoveryPointErrorKind

SetLocalConsolePasswordErrorKind

SetSMBGuestPasswordErrorKind

ShutdownGatewayErrorKind

StartAvailabilityMonitorTestErrorKind

StartGatewayErrorKind

UpdateAutomaticTapeCreationPolicyErrorKind

UpdateBandwidthRateLimitErrorKind

UpdateBandwidthRateLimitScheduleErrorKind

UpdateChapCredentialsErrorKind

UpdateFileSystemAssociationErrorKind

UpdateGatewayInformationErrorKind

UpdateGatewaySoftwareNowErrorKind

UpdateMaintenanceStartTimeErrorKind

UpdateNFSFileShareErrorKind

UpdateSMBFileShareErrorKind

UpdateSMBFileShareVisibilityErrorKind

UpdateSMBSecurityStrategyErrorKind

UpdateSnapshotScheduleErrorKind

UpdateVTLDeviceTypeErrorKind