Module aws_sdk_storagegateway::input

Modules

activate_gateway_input

See ActivateGatewayInput

add_cache_input

See AddCacheInput

add_tags_to_resource_input

See AddTagsToResourceInput

add_upload_buffer_input

See AddUploadBufferInput

add_working_storage_input

See AddWorkingStorageInput

assign_tape_pool_input

See AssignTapePoolInput

associate_file_system_input

See AssociateFileSystemInput

attach_volume_input

See AttachVolumeInput

cancel_archival_input

See CancelArchivalInput

cancel_retrieval_input

See CancelRetrievalInput

create_cachedi_scsi_volume_input

See CreateCachediScsiVolumeInput

create_nfs_file_share_input

See CreateNfsFileShareInput

create_smb_file_share_input

See CreateSmbFileShareInput

create_snapshot_from_volume_recovery_point_input

See CreateSnapshotFromVolumeRecoveryPointInput

create_snapshot_input

See CreateSnapshotInput

create_storedi_scsi_volume_input

See CreateStorediScsiVolumeInput

create_tape_pool_input

See CreateTapePoolInput

create_tape_with_barcode_input

See CreateTapeWithBarcodeInput

create_tapes_input

See CreateTapesInput

delete_automatic_tape_creation_policy_input

See DeleteAutomaticTapeCreationPolicyInput

delete_bandwidth_rate_limit_input

See DeleteBandwidthRateLimitInput

delete_chap_credentials_input

See DeleteChapCredentialsInput

delete_file_share_input

See DeleteFileShareInput

delete_gateway_input

See DeleteGatewayInput

delete_snapshot_schedule_input

See DeleteSnapshotScheduleInput

delete_tape_archive_input

See DeleteTapeArchiveInput

delete_tape_input

See DeleteTapeInput

delete_tape_pool_input

See DeleteTapePoolInput

delete_volume_input

See DeleteVolumeInput

describe_availability_monitor_test_input

See DescribeAvailabilityMonitorTestInput

describe_bandwidth_rate_limit_input

See DescribeBandwidthRateLimitInput

describe_bandwidth_rate_limit_schedule_input

See DescribeBandwidthRateLimitScheduleInput

describe_cache_input

See DescribeCacheInput

describe_cachedi_scsi_volumes_input

See DescribeCachediScsiVolumesInput

describe_chap_credentials_input

See DescribeChapCredentialsInput

describe_file_system_associations_input

See DescribeFileSystemAssociationsInput

describe_gateway_information_input

See DescribeGatewayInformationInput

describe_maintenance_start_time_input

See DescribeMaintenanceStartTimeInput

describe_nfs_file_shares_input

See DescribeNfsFileSharesInput

describe_smb_file_shares_input

See DescribeSmbFileSharesInput

describe_smb_settings_input

See DescribeSmbSettingsInput

describe_snapshot_schedule_input

See DescribeSnapshotScheduleInput

describe_storedi_scsi_volumes_input

See DescribeStorediScsiVolumesInput

describe_tape_archives_input

See DescribeTapeArchivesInput

describe_tape_recovery_points_input

See DescribeTapeRecoveryPointsInput

describe_tapes_input

See DescribeTapesInput

describe_upload_buffer_input

See DescribeUploadBufferInput

describe_vtl_devices_input

See DescribeVtlDevicesInput

describe_working_storage_input

See DescribeWorkingStorageInput

detach_volume_input

See DetachVolumeInput

disable_gateway_input

See DisableGatewayInput

disassociate_file_system_input

See DisassociateFileSystemInput

join_domain_input

See JoinDomainInput

list_automatic_tape_creation_policies_input

See ListAutomaticTapeCreationPoliciesInput

list_file_shares_input

See ListFileSharesInput

list_file_system_associations_input

See ListFileSystemAssociationsInput

list_gateways_input

See ListGatewaysInput

list_local_disks_input

See ListLocalDisksInput

list_tags_for_resource_input

See ListTagsForResourceInput

list_tape_pools_input

See ListTapePoolsInput

list_tapes_input

See ListTapesInput

list_volume_initiators_input

See ListVolumeInitiatorsInput

list_volume_recovery_points_input

See ListVolumeRecoveryPointsInput

list_volumes_input

See ListVolumesInput

notify_when_uploaded_input

See NotifyWhenUploadedInput

refresh_cache_input

See RefreshCacheInput

remove_tags_from_resource_input

See RemoveTagsFromResourceInput

reset_cache_input

See ResetCacheInput

retrieve_tape_archive_input

See RetrieveTapeArchiveInput

retrieve_tape_recovery_point_input

See RetrieveTapeRecoveryPointInput

set_local_console_password_input

See SetLocalConsolePasswordInput

set_smb_guest_password_input

See SetSmbGuestPasswordInput

shutdown_gateway_input

See ShutdownGatewayInput

start_availability_monitor_test_input

See StartAvailabilityMonitorTestInput

start_gateway_input

See StartGatewayInput

update_automatic_tape_creation_policy_input

See UpdateAutomaticTapeCreationPolicyInput

update_bandwidth_rate_limit_input

See UpdateBandwidthRateLimitInput

update_bandwidth_rate_limit_schedule_input

See UpdateBandwidthRateLimitScheduleInput

update_chap_credentials_input

See UpdateChapCredentialsInput

update_file_system_association_input

See UpdateFileSystemAssociationInput

update_gateway_information_input

See UpdateGatewayInformationInput

update_gateway_software_now_input

See UpdateGatewaySoftwareNowInput

update_maintenance_start_time_input

See UpdateMaintenanceStartTimeInput

update_nfs_file_share_input

See UpdateNfsFileShareInput

update_smb_file_share_input

See UpdateSmbFileShareInput

update_smb_file_share_visibility_input

See UpdateSmbFileShareVisibilityInput

update_smb_security_strategy_input

See UpdateSmbSecurityStrategyInput

update_snapshot_schedule_input

See UpdateSnapshotScheduleInput

update_vtl_device_type_input

See UpdateVtlDeviceTypeInput

Structs

ActivateGatewayInput

A JSON object containing one or more of the following fields:

AddCacheInput

AddTagsToResourceInput

AddTagsToResourceInput

AddUploadBufferInput

AddWorkingStorageInput

A JSON object containing one or more of the following fields:

AssignTapePoolInput

AssociateFileSystemInput

AttachVolumeInput

AttachVolumeInput

CancelArchivalInput

CancelArchivalInput

CancelRetrievalInput

CancelRetrievalInput

CreateCachediScsiVolumeInput

CreateNfsFileShareInput

CreateNFSFileShareInput

CreateSmbFileShareInput

CreateSMBFileShareInput

CreateSnapshotFromVolumeRecoveryPointInput

CreateSnapshotInput

A JSON object containing one or more of the following fields:

CreateStorediScsiVolumeInput

A JSON object containing one or more of the following fields:

CreateTapePoolInput

CreateTapeWithBarcodeInput

CreateTapeWithBarcodeInput

CreateTapesInput

CreateTapesInput

DeleteAutomaticTapeCreationPolicyInput

DeleteBandwidthRateLimitInput

A JSON object containing the following fields:

DeleteChapCredentialsInput

A JSON object containing one or more of the following fields:

DeleteFileShareInput

DeleteFileShareInput

DeleteGatewayInput

A JSON object containing the ID of the gateway to delete.

DeleteSnapshotScheduleInput

DeleteTapeArchiveInput

DeleteTapeArchiveInput

DeleteTapeInput

DeleteTapeInput

DeleteTapePoolInput

DeleteVolumeInput

A JSON object containing the DeleteVolumeInput$VolumeARN to

DescribeAvailabilityMonitorTestInput

DescribeBandwidthRateLimitInput

A JSON object containing the Amazon Resource Name (ARN) of the gateway.

DescribeBandwidthRateLimitScheduleInput

DescribeCacheInput

DescribeCachediScsiVolumesInput

DescribeChapCredentialsInput

A JSON object containing the Amazon Resource Name (ARN) of the iSCSI volume

DescribeFileSystemAssociationsInput

DescribeGatewayInformationInput

A JSON object containing the ID of the gateway.

DescribeMaintenanceStartTimeInput

A JSON object containing the Amazon Resource Name (ARN) of the gateway.

DescribeNfsFileSharesInput

DescribeNFSFileSharesInput

DescribeSmbFileSharesInput

DescribeSMBFileSharesInput

DescribeSmbSettingsInput

DescribeSnapshotScheduleInput

A JSON object containing the DescribeSnapshotScheduleInput$VolumeARN

DescribeStorediScsiVolumesInput

A JSON object containing a list of DescribeStorediSCSIVolumesInput$VolumeARNs.

DescribeTapeArchivesInput

DescribeTapeArchivesInput

DescribeTapeRecoveryPointsInput

DescribeTapeRecoveryPointsInput

DescribeTapesInput

DescribeTapesInput

DescribeUploadBufferInput

DescribeVtlDevicesInput

DescribeVTLDevicesInput

DescribeWorkingStorageInput

A JSON object containing the Amazon Resource Name (ARN) of the gateway.

DetachVolumeInput

AttachVolumeInput

DisableGatewayInput

DisableGatewayInput

DisassociateFileSystemInput

JoinDomainInput

JoinDomainInput

ListAutomaticTapeCreationPoliciesInput

ListFileSharesInput

ListFileShareInput

ListFileSystemAssociationsInput

ListGatewaysInput

A JSON object containing zero or more of the following fields:

ListLocalDisksInput

A JSON object containing the Amazon Resource Name (ARN) of the gateway.

ListTagsForResourceInput

ListTagsForResourceInput

ListTapePoolsInput

ListTapesInput

A JSON object that contains one or more of the following fields:

ListVolumeInitiatorsInput

ListVolumeInitiatorsInput

ListVolumeRecoveryPointsInput

ListVolumesInput

A JSON object that contains one or more of the following fields:

NotifyWhenUploadedInput

RefreshCacheInput

RefreshCacheInput

RemoveTagsFromResourceInput

RemoveTagsFromResourceInput

ResetCacheInput

RetrieveTapeArchiveInput

RetrieveTapeArchiveInput

RetrieveTapeRecoveryPointInput

RetrieveTapeRecoveryPointInput

SetLocalConsolePasswordInput

SetLocalConsolePasswordInput

SetSmbGuestPasswordInput

SetSMBGuestPasswordInput

ShutdownGatewayInput

A JSON object containing the Amazon Resource Name (ARN) of the gateway to shut

StartAvailabilityMonitorTestInput

StartGatewayInput

A JSON object containing the Amazon Resource Name (ARN) of the gateway to start.

UpdateAutomaticTapeCreationPolicyInput

UpdateBandwidthRateLimitInput

A JSON object containing one or more of the following fields:

UpdateBandwidthRateLimitScheduleInput

UpdateChapCredentialsInput

A JSON object containing one or more of the following fields:

UpdateFileSystemAssociationInput

UpdateGatewayInformationInput

UpdateGatewaySoftwareNowInput

A JSON object containing the Amazon Resource Name (ARN) of the gateway to update.

UpdateMaintenanceStartTimeInput

A JSON object containing the following fields:

UpdateNfsFileShareInput

UpdateNFSFileShareInput

UpdateSmbFileShareInput

UpdateSMBFileShareInput

UpdateSmbFileShareVisibilityInput

UpdateSmbSecurityStrategyInput

UpdateSnapshotScheduleInput

A JSON object containing one or more of the following fields:

UpdateVtlDeviceTypeInput