Struct aws_sdk_storagegateway::input::UpdateSmbFileShareInput

#[non_exhaustive]
pub struct UpdateSmbFileShareInput {

    pub file_share_arn: Option<String>,
    pub kms_encrypted: Option<bool>,
    pub kms_key: Option<String>,
    pub default_storage_class: Option<String>,
    pub object_acl: Option<ObjectAcl>,
    pub read_only: Option<bool>,
    pub guess_mime_type_enabled: Option<bool>,
    pub requester_pays: Option<bool>,
    pub smbacl_enabled: Option<bool>,
    pub access_based_enumeration: Option<bool>,
    pub admin_user_list: Option<Vec<String>>,
    pub valid_user_list: Option<Vec<String>>,
    pub invalid_user_list: Option<Vec<String>>,
    pub audit_destination_arn: Option<String>,
    pub case_sensitivity: Option<CaseSensitivity>,
    pub file_share_name: Option<String>,
    pub cache_attributes: Option<CacheAttributes>,
    pub notification_policy: Option<String>,
    pub oplocks_enabled: Option<bool>,

}

UpdateSMBFileShareInput

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

file_share_arn: Option<String>

The Amazon Resource Name (ARN) of the SMB file share that you want to update.

kms_encrypted: Option<bool>

Set to true to use Amazon S3 server-side encryption with your own KMS key, or false to use a key managed by Amazon S3. Optional.

Valid Values: true | false

kms_key: Option<String>

The Amazon Resource Name (ARN) of a symmetric customer master key (CMK) used for Amazon S3 server-side encryption. Storage Gateway does not support asymmetric CMKs. This value can only be set when KMSEncrypted is true. Optional.

default_storage_class: Option<String>

The default storage class for objects put into an Amazon S3 bucket by the S3 File Gateway. The default value is S3_INTELLIGENT_TIERING. Optional.

Valid Values: S3_STANDARD | S3_INTELLIGENT_TIERING | S3_STANDARD_IA | S3_ONEZONE_IA

object_acl: Option<ObjectAcl>

A value that sets the access control list (ACL) permission for objects in the S3 bucket that a S3 File Gateway puts objects into. The default value is private.

read_only: Option<bool>

A value that sets the write status of a file share. Set this value to true to set write status to read-only, otherwise set to false.

Valid Values: true | false

guess_mime_type_enabled: Option<bool>

A value that enables guessing of the MIME type for uploaded objects based on file extensions. Set this value to true to enable MIME type guessing, otherwise set to false. The default value is true.

Valid Values: true | false

requester_pays: Option<bool>

A value that sets who pays the cost of the request and the cost associated with data download from the S3 bucket. If this value is set to true, the requester pays the costs; otherwise, the S3 bucket owner pays. However, the S3 bucket owner always pays the cost of storing data.

RequesterPays is a configuration for the S3 bucket that backs the file share, so make sure that the configuration on the file share is the same as the S3 bucket configuration.

Valid Values: true | false

smbacl_enabled: Option<bool>

Set this value to true to enable access control list (ACL) on the SMB file share. Set it to false to map file and directory permissions to the POSIX permissions.

For more information, see Using Microsoft Windows ACLs to control access to an SMB file share in the Storage Gateway User Guide.

Valid Values: true | false

access_based_enumeration: Option<bool>

The files and folders on this share will only be visible to users with read access.

admin_user_list: Option<Vec<String>>

A list of users or groups in the Active Directory that have administrator rights to the file share. A group must be prefixed with the @ character. Acceptable formats include: DOMAIN\User1, user1, @group1, and @DOMAIN\group1. Can only be set if Authentication is set to ActiveDirectory.

valid_user_list: Option<Vec<String>>

A list of users or groups in the Active Directory that are allowed to access the file share. A group must be prefixed with the @ character. Acceptable formats include: DOMAIN\User1, user1, @group1, and @DOMAIN\group1. Can only be set if Authentication is set to ActiveDirectory.

invalid_user_list: Option<Vec<String>>

A list of users or groups in the Active Directory that are not allowed to access the file share. A group must be prefixed with the @ character. Acceptable formats include: DOMAIN\User1, user1, @group1, and @DOMAIN\group1. Can only be set if Authentication is set to ActiveDirectory.

audit_destination_arn: Option<String>

The Amazon Resource Name (ARN) of the storage used for audit logs.

case_sensitivity: Option<CaseSensitivity>

The case of an object name in an Amazon S3 bucket. For ClientSpecified, the client determines the case sensitivity. For CaseSensitive, the gateway determines the case sensitivity. The default value is ClientSpecified.

file_share_name: Option<String>

The name of the file share. Optional.

FileShareName must be set if an S3 prefix name is set in LocationARN.

cache_attributes: Option<CacheAttributes>

Specifies refresh cache information for the file share.

notification_policy: Option<String>

The notification policy of the file share. SettlingTimeInSeconds controls the number of seconds to wait after the last point in time a client wrote to a file before generating an ObjectUploaded notification. Because clients can make many small writes to files, it's best to set this parameter for as long as possible to avoid generating multiple notifications for the same file in a small time period.

SettlingTimeInSeconds has no effect on the timing of the object uploading to Amazon S3, only the timing of the notification.

The following example sets NotificationPolicy on with SettlingTimeInSeconds set to 60.

{\"Upload\": {\"SettlingTimeInSeconds\": 60}}

The following example sets NotificationPolicy off.

{}

oplocks_enabled: Option<bool>

Specifies whether opportunistic locking is enabled for the SMB file share.

Enabling opportunistic locking on case-sensitive shares is not recommended for workloads that involve access to files with the same name in different case.

Valid Values: true | false

Implementations

impl UpdateSmbFileShareInput

pub fn make_operation(
&self,
_config: &Config
) -> Result<Operation<UpdateSMBFileShare, AwsErrorRetryPolicy>, BuildError>

Consumes the builder and constructs an Operation<UpdateSMBFileShare>

pub fn builder() -> Builder

Creates a new builder-style object to manufacture UpdateSmbFileShareInput