Module aws_sdk_storagegateway::model

Modules

automatic_tape_creation_policy_info

See AutomaticTapeCreationPolicyInfo

automatic_tape_creation_rule

See AutomaticTapeCreationRule

bandwidth_rate_limit_interval

See BandwidthRateLimitInterval

cache_attributes

See CacheAttributes

cachedi_scsi_volume

See CachediScsiVolume

chap_info

See ChapInfo

devicei_scsi_attributes

See DeviceiScsiAttributes

disk

See Disk

endpoint_network_configuration

See EndpointNetworkConfiguration

file_share_info

See FileShareInfo

file_system_association_info

See FileSystemAssociationInfo

file_system_association_summary

See FileSystemAssociationSummary

gateway_info

See GatewayInfo

network_interface

See NetworkInterface

nfs_file_share_defaults

See NfsFileShareDefaults

nfs_file_share_info

See NfsFileShareInfo

pool_info

See PoolInfo

smb_file_share_info

See SmbFileShareInfo

storage_gateway_error

See StorageGatewayError

storedi_scsi_volume

See StorediScsiVolume

tag

See Tag

tape

See Tape

tape_archive

See TapeArchive

tape_info

See TapeInfo

tape_recovery_point_info

See TapeRecoveryPointInfo

volume_info

See VolumeInfo

volume_recovery_point_info

See VolumeRecoveryPointInfo

volumei_scsi_attributes

See VolumeiScsiAttributes

vtl_device

See VtlDevice

Structs

AutomaticTapeCreationPolicyInfo

Information about the gateway's automatic tape creation policies, including the

AutomaticTapeCreationRule

An automatic tape creation policy consists of automatic tape creation rules where each

BandwidthRateLimitInterval

CacheAttributes

The refresh cache information for the file share or FSx file systems.

CachediScsiVolume

Describes an iSCSI cached volume.

ChapInfo

Describes Challenge-Handshake Authentication Protocol (CHAP) information that supports

DeviceiScsiAttributes

Lists iSCSI information about a VTL device.

Disk

Represents a gateway's local disk.

EndpointNetworkConfiguration

Specifies network configuration information for the gateway associated with the

FileShareInfo

Describes a file share. Only supported S3 File Gateway.

FileSystemAssociationInfo

Describes the object returned by DescribeFileSystemAssociations that

FileSystemAssociationSummary

Gets the summary returned by ListFileSystemAssociation, which is a summary

GatewayInfo

Describes a gateway object.

NetworkInterface

Describes a gateway's network interface.

NfsFileShareDefaults

Describes Network File System (NFS) file share default values. Files and folders stored

NfsFileShareInfo

The Unix file permissions and ownership information assigned, by default, to native S3

PoolInfo

Describes a custom tape pool.

SmbFileShareInfo

The Windows file permissions and ownership information assigned, by default, to native

StorageGatewayError

Provides additional information about an error that was returned by the service. See the

StorediScsiVolume

Describes an iSCSI stored volume.

Tag

A key-value pair that helps you manage, filter, and search for your resource. Allowed

Tape

Describes a virtual tape object.

TapeArchive

Represents a virtual tape that is archived in the virtual tape shelf (VTS).

TapeInfo

Describes a virtual tape.

TapeRecoveryPointInfo

Describes a recovery point.

VolumeInfo

Describes a storage volume object.

VolumeRecoveryPointInfo

Describes a storage volume recovery point object.

VolumeiScsiAttributes

Lists iSCSI information about a volume.

VtlDevice

Represents a device object associated with a tape gateway.

Enums

ActiveDirectoryStatus

AvailabilityMonitorTestStatus

CaseSensitivity

ErrorCode

FileShareType

The type of the file share.

GatewayCapacity

HostEnvironment

ObjectAcl

A value that sets the access control list (ACL) permission for objects in the S3 bucket

PoolStatus

RetentionLockType

SmbSecurityStrategy

TapeStorageClass