Module aws_sdk_storagegateway::operation

Structs

ActivateGateway

Activates the gateway you previously deployed on your host. In the activation process,

AddCache

Configures one or more gateway local disks as cache for a gateway. This operation is

AddTagsToResource

Adds one or more tags to the specified resource. You use tags to add metadata to

AddUploadBuffer

Configures one or more gateway local disks as upload buffer for a specified gateway.

AddWorkingStorage

Configures one or more gateway local disks as working storage for a gateway. This

AssignTapePool

Assigns a tape to a tape pool for archiving. The tape assigned to a pool is archived in

AssociateFileSystem

Associate an Amazon FSx file system with the FSx File Gateway. After the

AttachVolume

Connects a volume to an iSCSI connection and then attaches the volume to the specified

CancelArchival

Cancels archiving of a virtual tape to the virtual tape shelf (VTS) after the archiving

CancelRetrieval

Cancels retrieval of a virtual tape from the virtual tape shelf (VTS) to a gateway after

CreateCachediSCSIVolume

Creates a cached volume on a specified cached volume gateway. This operation is only

CreateNFSFileShare

Creates a Network File System (NFS) file share on an existing S3 File Gateway. In Storage

CreateSMBFileShare

Creates a Server Message Block (SMB) file share on an existing S3 File Gateway. In Storage

CreateSnapshot

Initiates a snapshot of a volume.

CreateSnapshotFromVolumeRecoveryPoint

Initiates a snapshot of a gateway from a volume recovery point. This operation is only

CreateStorediSCSIVolume

Creates a volume on a specified gateway. This operation is only supported in the stored

CreateTapePool

Creates a new custom tape pool. You can use custom tape pool to enable tape retention

CreateTapeWithBarcode

Creates a virtual tape by using your own barcode. You write data to the virtual tape and

CreateTapes

Creates one or more virtual tapes. You write data to the virtual tapes and then archive

DeleteAutomaticTapeCreationPolicy

Deletes the automatic tape creation policy of a gateway. If you delete this policy, new

DeleteBandwidthRateLimit

Deletes the bandwidth rate limits of a gateway. You can delete either the upload and

DeleteChapCredentials

Deletes Challenge-Handshake Authentication Protocol (CHAP) credentials for a specified

DeleteFileShare

Deletes a file share from an S3 File Gateway. This operation is only supported for S3 File

DeleteGateway

Deletes a gateway. To specify which gateway to delete, use the Amazon Resource Name

DeleteSnapshotSchedule

Deletes a snapshot of a volume.

DeleteTape

Deletes the specified virtual tape. This operation is only supported in the tape gateway

DeleteTapeArchive

Deletes the specified virtual tape from the virtual tape shelf (VTS). This operation is

DeleteTapePool

Delete a custom tape pool. A custom tape pool can only be deleted if there are no tapes

DeleteVolume

Deletes the specified storage volume that you previously created using the CreateCachediSCSIVolume or CreateStorediSCSIVolume API.

DescribeAvailabilityMonitorTest

Returns information about the most recent high availability monitoring test that was

DescribeBandwidthRateLimit

Returns the bandwidth rate limits of a gateway. By default, these limits are not set,

DescribeBandwidthRateLimitSchedule

DescribeCache

Returns information about the cache of a gateway. This operation is only supported in

DescribeCachediSCSIVolumes

Returns a description of the gateway volumes specified in the request. This operation is

DescribeChapCredentials

Returns an array of Challenge-Handshake Authentication Protocol (CHAP) credentials

DescribeFileSystemAssociations

Gets the file system association information. This operation is only supported for

DescribeGatewayInformation

Returns metadata about a gateway such as its name, network interfaces, configured time

DescribeMaintenanceStartTime

Returns your gateway's weekly maintenance start time including the day and time of

DescribeNFSFileShares

Gets a description for one or more Network File System (NFS) file shares from an S3 File

DescribeSMBFileShares

Gets a description for one or more Server Message Block (SMB) file shares from a S3 File

DescribeSMBSettings

Gets a description of a Server Message Block (SMB) file share settings from a file

DescribeSnapshotSchedule

Describes the snapshot schedule for the specified gateway volume. The snapshot schedule

DescribeStorediSCSIVolumes

Returns the description of the gateway volumes specified in the request. The list of

DescribeTapeArchives

Returns a description of specified virtual tapes in the virtual tape shelf (VTS). This

DescribeTapeRecoveryPoints

Returns a list of virtual tape recovery points that are available for the specified tape

DescribeTapes

Returns a description of the specified Amazon Resource Name (ARN) of virtual tapes. If a

DescribeUploadBuffer

Returns information about the upload buffer of a gateway. This operation is supported

DescribeVTLDevices

Returns a description of virtual tape library (VTL) devices for the specified tape

DescribeWorkingStorage

Returns information about the working storage of a gateway. This operation is only

DetachVolume

Disconnects a volume from an iSCSI connection and then detaches the volume from the

DisableGateway

Disables a tape gateway when the gateway is no longer functioning. For example, if your

DisassociateFileSystem

Disassociates an Amazon FSx file system from the specified gateway. After the

JoinDomain

Adds a file gateway to an Active Directory domain. This operation is only supported for

ListAutomaticTapeCreationPolicies

Lists the automatic tape creation policies for a gateway. If there are no automatic tape

ListFileShares

Gets a list of the file shares for a specific S3 File Gateway, or the list of file shares

ListFileSystemAssociations

Gets a list of FileSystemAssociationSummary objects. Each object contains a

ListGateways

Lists gateways owned by an account in an Region specified in the request. The

ListLocalDisks

Returns a list of the gateway's local disks. To specify which gateway to describe,

ListTagsForResource

Lists the tags that have been added to the specified resource. This operation is

ListTapePools

Lists custom tape pools. You specify custom tape pools to list by specifying one or more

ListTapes

Lists virtual tapes in your virtual tape library (VTL) and your virtual tape shelf

ListVolumeInitiators

Lists iSCSI initiators that are connected to a volume. You can use this operation to

ListVolumeRecoveryPoints

Lists the recovery points for a specified gateway. This operation is only supported in

ListVolumes

Lists the iSCSI stored volumes of a gateway. Results are sorted by volume ARN. The

NotifyWhenUploaded

Sends you notification through CloudWatch Events when all files written to your file

RefreshCache

Refreshes the cached inventory of objects for the specified file share. This operation

RemoveTagsFromResource

Removes one or more tags from the specified resource. This operation is supported in

ResetCache

Resets all cache disks that have encountered an error and makes the disks available for

RetrieveTapeArchive

Retrieves an archived virtual tape from the virtual tape shelf (VTS) to a tape gateway.

RetrieveTapeRecoveryPoint

Retrieves the recovery point for the specified virtual tape. This operation is only

SetLocalConsolePassword

Sets the password for your VM local console. When you log in to the local console for

SetSMBGuestPassword

Sets the password for the guest user smbguest. The smbguest

ShutdownGateway

Shuts down a gateway. To specify which gateway to shut down, use the Amazon Resource

StartAvailabilityMonitorTest

Start a test that verifies that the specified gateway is configured for High

StartGateway

Starts a gateway that you previously shut down (see ShutdownGateway).

UpdateAutomaticTapeCreationPolicy

Updates the automatic tape creation policy of a gateway. Use this to update the policy

UpdateBandwidthRateLimit

Updates the bandwidth rate limits of a gateway. You can update both the upload and

UpdateBandwidthRateLimitSchedule

UpdateChapCredentials

Updates the Challenge-Handshake Authentication Protocol (CHAP) credentials for a

UpdateFileSystemAssociation

Updates a file system association. This operation is only supported in the FSx File Gateways.

UpdateGatewayInformation

Updates a gateway's metadata, which includes the gateway's name and time zone.

UpdateGatewaySoftwareNow

Updates the gateway virtual machine (VM) software. The request immediately triggers the

UpdateMaintenanceStartTime

Updates a gateway's weekly maintenance start time information, including day and

UpdateNFSFileShare

Updates a Network File System (NFS) file share. This operation is only supported in S3 File Gateways.

UpdateSMBFileShare

Updates a Server Message Block (SMB) file share. This operation is only supported for

UpdateSMBFileShareVisibility

Controls whether the shares on an S3 File Gateway are visible in a net view or browse

UpdateSMBSecurityStrategy

Updates the SMB security strategy on a file gateway. This action is only supported in

UpdateSnapshotSchedule

Updates a snapshot schedule configured for a gateway volume. This operation is only

UpdateVTLDeviceType

Updates the type of medium changer in a tape gateway. When you activate a tape gateway,